Import 2.1.121pre1

+++ drivers/block/floppy.c
@@ -233,7 +233,7 @@ static inline void fallback_on_nodma_alloc(char **addr, size_t l)
 
 /* End dma memory related stuff */
 
-static unsigned int fake_change = 0;
+static unsigned long fake_change = 0;
 static int initialising=1;
 
 static inline int TYPE(kdev_t x) {
